requirements.txt
anyio==3.7.0
black==23.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
distlib==0.3.6
exceptiongroup==1.1.1
Fenv==0.0.12.4
filelock==3.12.2
flet==0.7.4
flet-core==0.7.4
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.3
python-dotenv==1.0.0
repath==0.9.0
requests==2.31.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
tomli==2.0.1
urllib3==2.0.3
virtualenv==20.23.0
watchdog==2.3.1
websocket-client==1.5.3
websockets==10.4
